Time: 3 ms (37.39%), Space: 17 MB (85.9%) - LeetHub

+++ 0150-evaluate-reverse-polish-notation/0150-evaluate-reverse-polish-notation.cpp
@@ -8,14 +8,14 @@ public:
             return a-b;
         }
         if(token == "*"){
-            return a*b;
+            return long(a)*long(b);
         }
         if(token == "/"){
             return a/b;
         }
         return -1;
     }
-    
+
     int evalRPN(vector<string>& tokens) {
         stack<int> st;
